% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{fetch_floodpulse}
\alias{fetch_floodpulse}
\title{Fetch Sentinel-1 GRD Data from Google Earth Engine}
\usage{
fetch_floodpulse(aoi, start_date, end_date)
}
\arguments{
\item{aoi}{An sf object defining the area of interest (or a path to a shapefile)}

\item{start_date}{Start date (YYYYMMDD)}

\item{end_date}{End date (YYYYMMDD)}
}
\value{
Asset ID of the exported Sentinel-1 image in GEE
}
\description{
Fetch Sentinel-1 GRD Data from Google Earth Engine
}
