% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_floodpulse}
\alias{preview_floodpulse}
\title{Preview Sentinel-1 COG Metadata}
\usage{
preview_floodpulse(cog_path)
}
\arguments{
\item{cog_path}{Path to downloaded COG file}
}
\value{
List of metadata
}
\description{
Preview Sentinel-1 COG Metadata
}
