% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{start_floodpulse}
\alias{start_floodpulse}
\title{Start FloodPulseRNew GUI}
\usage{
start_floodpulse()
}
\description{
Launches a Shiny application for monitoring floods using Sentinel-1 imagery from Google Earth Engine.
Requires a Google Earth Engine Service Account Key (SaK) file for initialization.
}
